test(deprecate): split tests to mirror modules
- rename `test_property.py` -> `test__properties.py` and `test_call_plan.py` -> `test__dispatch.py`
- split the mixed `test_deprecation.py` by target module into `test__dispatch.py` (target resolution + call-plan), `test_messages.py` (templates + emitters), `test_callables.py` (packing decorators), and `test_docstring_inject.py` (docstring helpers)
- repoint imports at the real modules and update the stacklevel filename assertions to the renamed files; every test case preserved (1492 pass)

---
